extract connectionStore to a separate file
