+++ deploy/azure/bicep/main.bicep
@@ -8,7 +8,7 @@ param containerAppEnvironmentName string = ''
 param containerAppName string = 'mongo-mcp-server-app'
 
 @description('Docker image to deploy')
-param containerImage string = 'mongodb/mongodb-mcp-server:1.15.0-alpha.1'
+param containerImage string = 'mongodb/mongodb-mcp-server:1.15.0-alpha.2'
 
 @description('Container CPU (vCPU) as string. Allowed: 0.25 - 2.0 in 0.25 increments')
 @allowed([
