+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "mongodb-mcp-server",
   "description": "MongoDB Model Context Protocol Server",
-  "version": "1.15.0-alpha.1",
+  "version": "1.15.0-alpha.2",
   "type": "module",
   "packageManager": "pnpm@11.5.1",
   "mcpName": "io.github.mongodb-js/mongodb-mcp-server",
