+++ packages/metrics/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@mongodb-js/mcp-metrics",
   "description": "Prometheus metrics helpers for MongoDB MCP servers",
-  "version": "1.15.0-alpha.1",
+  "version": "1.15.0-alpha.2",
   "type": "module",
   "license": "Apache-2.0",
   "author": "MongoDB <info@mongodb.com>",
