+++ packages/types/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@mongodb-js/mcp-types",
   "description": "Shared TypeScript types and interfaces for MongoDB MCP server packages",
-  "version": "1.15.0-alpha.1",
+  "version": "1.15.0-alpha.2",
   "type": "module",
   "license": "Apache-2.0",
   "author": "MongoDB <info@mongodb.com>",
