+++ src/common/packageInfo.ts
@@ -4,7 +4,7 @@ export const packageInfo: {
     mcpServerName: string;
     engines: { node: string };
 } = {
-    version: "1.15.0-alpha.1",
+    version: "1.15.0-alpha.2",
     mcpServerName: "MongoDB MCP Server",
     engines: {
         node: "^20.19.0 || ^22.13.0 || >= 24.0.0",
