+++ src/lib.ts
@@ -75,17 +75,19 @@ export {
     type OIDCConnectionAuthType,
 } from "./common/connectionManager.js";
 export {
-    MCPConnectionStore,
     ConnectionEntry,
     PRECONFIGURED_CONNECTION_ID,
     atlasClusterSlug,
     type ConnectionRegistry,
-    type ConnectionStoreOptions,
     type ConnectionSource,
     type CreateConnectionEntryOptions,
     type CreateConnectionOptions,
-    type CreateConnectionManagerFn,
 } from "./common/connectionRegistry.js";
+export {
+    MCPConnectionStore,
+    type ConnectionStoreOptions,
+    type CreateConnectionManagerFn,
+} from "./common/connectionStore.js";
 export {
     connectionErrorHandler,
     type ConnectionErrorHandler,
