+++ src/web.ts
@@ -52,17 +52,19 @@ export type {
 } from "./common/connectionManager.js";
 export { ConnectionManager, ConnectionStateConnected } from "./common/connectionManager.js";
 export {
-    MCPConnectionStore,
     ConnectionEntry,
     PRECONFIGURED_CONNECTION_ID,
     atlasClusterSlug,
     type ConnectionRegistry,
-    type ConnectionStoreOptions,
     type ConnectionSource,
     type CreateConnectionEntryOptions,
     type CreateConnectionOptions,
-    type CreateConnectionManagerFn,
 } from "./common/connectionRegistry.js";
+export {
+    MCPConnectionStore,
+    type ConnectionStoreOptions,
+    type CreateConnectionManagerFn,
+} from "./common/connectionStore.js";
 export {
     ExportsManager,
     type AvailableExport,
