+++ tests/integration/helpers.ts
@@ -10,7 +10,8 @@ import { type UserConfig } from "../../src/common/config/userConfig.js";
 import { ResourceUpdatedNotificationSchema } from "@modelcontextprotocol/sdk/types.js";
 import { afterAll, afterEach, beforeAll, describe, expect, it, vi } from "vitest";
 import type { ConnectionManager, ConnectionState } from "../../src/common/connectionManager.js";
-import { MCPConnectionStore, type ConnectionEntry } from "../../src/common/connectionRegistry.js";
+import { MCPConnectionStore } from "../../src/common/connectionStore.js";
+import type { ConnectionEntry } from "../../src/common/connectionRegistry.js";
 import { DeviceId } from "../../src/helpers/deviceId.js";
 import { connectionErrorHandler } from "../../src/common/connectionErrorHandler.js";
 import { Keychain } from "../../src/common/keychain.js";
