+++ tests/integration/server.test.ts
@@ -1,4 +1,4 @@
-import { MCPConnectionStore } from "../../src/common/connectionRegistry.js";
+import { MCPConnectionStore } from "../../src/common/connectionStore.js";
 import { ExportsManager } from "../../src/common/exportsManager.js";
 import { CompositeLogger } from "../../src/common/logging/index.js";
 import { DeviceId } from "../../src/helpers/deviceId.js";
