+++ tests/unit/common/session.test.ts
@@ -2,7 +2,8 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { Session } from "../../../src/common/session.js";
 import { CompositeLogger } from "../../../src/common/logging/index.js";
 import { ExportsManager } from "../../../src/common/exportsManager.js";
-import { MCPConnectionStore, type ConnectionRegistry } from "../../../src/common/connectionRegistry.js";
+import { MCPConnectionStore } from "../../../src/common/connectionStore.js";
+import type { ConnectionRegistry } from "../../../src/common/connectionRegistry.js";
 import { DeviceId } from "../../../src/helpers/deviceId.js";
 import { Keychain } from "../../../src/common/keychain.js";
 import { defaultTestConfig } from "../../integration/helpers.js";
