+++ tests/unit/resources/common/debug.test.ts
@@ -2,11 +2,8 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { DebugResource } from "../../../../src/resources/common/debug.js";
 import { Session } from "../../../../src/common/session.js";
 import { CompositeLogger } from "../../../../src/common/logging/index.js";
-import {
-    MCPConnectionStore,
-    PRECONFIGURED_CONNECTION_ID,
-    type ConnectionRegistry,
-} from "../../../../src/common/connectionRegistry.js";
+import { PRECONFIGURED_CONNECTION_ID, type ConnectionRegistry } from "../../../../src/common/connectionRegistry.js";
+import { MCPConnectionStore } from "../../../../src/common/connectionStore.js";
 import { ExportsManager } from "../../../../src/common/exportsManager.js";
 import { DeviceId } from "../../../../src/helpers/deviceId.js";
 import { Keychain } from "../../../../src/common/keychain.js";
