+++ tests/unit/tools/atlas/update/pauseResumeCluster.test.ts
@@ -12,7 +12,8 @@ import type { Elicitation } from "../../../../../src/elicitation.js";
 import { CompositeLogger } from "../../../../../src/common/logging/index.js";
 import type { ApiClient } from "../../../../../src/common/atlas/apiClient.js";
 import type { AtlasClusterConnectionInfo } from "../../../../../src/common/connectionInfo.js";
-import { MCPConnectionStore, type ConnectionRegistry } from "../../../../../src/common/connectionRegistry.js";
+import { MCPConnectionStore } from "../../../../../src/common/connectionStore.js";
+import type { ConnectionRegistry } from "../../../../../src/common/connectionRegistry.js";
 import { DeviceId } from "../../../../../src/helpers/deviceId.js";
 import { UIRegistry } from "../../../../../src/ui/registry/index.js";
 import { MockMetrics } from "../../../mocks/metrics.js";
